added fcfs

+++ ProjectOS2/Process.cs
@@ -11,8 +11,8 @@ namespace ProjectOS2
         public string name { get; set; }
 
        // public static string schedulerType { get; set; }
-        public decimal arrivalTime { get; set; }
-        public decimal processTime { get; set; }  
+        public int arrivalTime { get; set; }
+        public int processTime { get; set; }  
         public int priority { get; set; }
        
         public object this[string propertyName]
